zapatillas fix save
