+++ src/api/commands.py
@@ -1,6 +1,6 @@
 
 import click
-from api.models import db, User
+from .models import db, User
 
 """
 In this file, you can add as many commands as you want using the @app.cli.command decorator
