--- src/api/routes.py
@@ -1,22 +1,0 @@
-"""
-This module takes care of starting the API Server, Loading the DB and Adding the endpoints
-"""
-from flask import Flask, request, jsonify, url_for, Blueprint
-from api.models import db, User
-from api.utils import generate_sitemap, APIException
-from flask_cors import CORS
-
-api = Blueprint('api', __name__)
-
-# Allow CORS requests to this API
-CORS(api)
-
-
-@api.route('/hello', methods=['POST', 'GET'])
-def handle_hello():
-
-    response_body = {
-        "message": "Antes de hacer merge o pull request leer bien cada una de las acciones para no joder el proyecto"
-    }
-
-    return jsonify(response_body), 200
