+++ src/app.py
@@ -5,13 +5,12 @@ import os
 from flask import Flask, request, jsonify, url_for, send_from_directory
 from flask_migrate import Migrate
 from flask_swagger import swagger
-from api.utils import APIException, generate_sitemap
-from api.models import db
-from api.routes import api
-from api.admin import setup_admin
-from api.commands import setup_commands
-from routes.zapatillas import zapatillas_bp
-from routes.usuario import usuario_bp
+from .api.utils import APIException, generate_sitemap
+from .api.models import db
+from .api.admin import setup_admin
+from .api.commands import setup_commands
+from .api.routes.zapatillas import zapatillas_bp
+from .api.routes.usuario import usuario_bp
 from flask_jwt_extended import JWTManager
 # from models import Person
 
@@ -38,7 +37,6 @@ setup_admin(app)
 setup_commands(app)
 
 # Add all endpoints form the API with a "api" prefix
-app.register_blueprint(api, url_prefix='/api')
 app.register_blueprint(zapatillas_bp)
 app.register_blueprint(usuario_bp)
 # Handle/serialize errors like a JSON object
